Реорганизация работы с пользовательскими функциями
