--- helpers.php
@@ -1,295 +1,0 @@
-<?php
-/**
- * Проверяет переданную дату на соответствие формату 'ГГГГ-ММ-ДД'
- *
- * Примеры использования:
- * is_date_valid('2019-01-01'); // true
- * is_date_valid('2016-02-29'); // true
- * is_date_valid('2019-04-31'); // false
- * is_date_valid('10.10.2010'); // false
- * is_date_valid('10/10/2010'); // false
- *
- * @param string $date Дата в виде строки
- *
- * @return bool true при совпадении с форматом 'ГГГГ-ММ-ДД', иначе false
- */
-function is_date_valid(string $date) : bool {
-    $format_to_check = 'Y-m-d';
-    $dateTimeObj = date_create_from_format($format_to_check, $date);
-
-    return $dateTimeObj !== false && array_sum(date_get_last_errors()) === 0;
-}
-
-/**
- * Создает подготовленное выражение на основе готового SQL запроса и переданных данных
- *
- * @param $link mysqli Ресурс соединения
- * @param $sql string SQL запрос с плейсхолдерами вместо значений
- * @param array $data Данные для вставки на место плейсхолдеров
- *
- * @return mysqli_stmt Подготовленное выражение
- */
-function db_get_prepare_stmt($link, $sql, $data = []) {
-    $stmt = mysqli_prepare($link, $sql);
-
-    if ($stmt === false) {
-        $errorMsg = 'Не удалось инициализировать подготовленное выражение: ' . mysqli_error($link);
-        die($errorMsg);
-    }
-
-    if ($data) {
-        $types = '';
-        $stmt_data = [];
-
-        foreach ($data as $value) {
-            $type = 's';
-
-            if (is_int($value)) {
-                $type = 'i';
-            }
-            else if (is_string($value)) {
-                $type = 's';
-            }
-            else if (is_double($value)) {
-                $type = 'd';
-            }
-
-            if ($type) {
-                $types .= $type;
-                $stmt_data[] = $value;
-            }
-        }
-
-        $values = array_merge([$stmt, $types], $stmt_data);
-
-        $func = 'mysqli_stmt_bind_param';
-        $func(...$values);
-
-        if (mysqli_errno($link) > 0) {
-            $errorMsg = 'Не удалось связать подготовленное выражение с параметрами: ' . mysqli_error($link);
-            die($errorMsg);
-        }
-    }
-
-    return $stmt;
-}
-
-/**
- * Возвращает корректную форму множественного числа
- * Ограничения: только для целых чисел
- *
- * Пример использования:
- * $remaining_minutes = 5;
- * echo "Я поставил таймер на {$remaining_minutes} " .
- *     get_noun_plural_form(
- *         $remaining_minutes,
- *         'минута',
- *         'минуты',
- *         'минут'
- *     );
- * Результат: "Я поставил таймер на 5 минут"
- *
- * @param int $number Число, по которому вычисляем форму множественного числа
- * @param string $one Форма единственного числа: яблоко, час, минута
- * @param string $two Форма множественного числа для 2, 3, 4: яблока, часа, минуты
- * @param string $many Форма множественного числа для остальных чисел
- *
- * @return string Рассчитанная форма множественнго числа
- */
-function get_noun_plural_form (int $number, string $one, string $two, string $many): string
-{
-    $number = (int) $number;
-    $mod10 = $number % 10;
-    $mod100 = $number % 100;
-
-    switch (true) {
-        case ($mod100 >= 11 && $mod100 <= 20):
-            return $many;
-
-        case ($mod10 > 5):
-            return $many;
-
-        case ($mod10 === 1):
-            return $one;
-
-        case ($mod10 >= 2 && $mod10 <= 4):
-            return $two;
-
-        default:
-            return $many;
-    }
-}
-
-/**
- * Подключает шаблон, передает туда данные и возвращает итоговый HTML контент
- * @param string $name Путь к файлу шаблона относительно папки templates
- * @param array $data Ассоциативный массив с данными для шаблона
- * @return string Итоговый HTML
- */
-function include_template($name, array $data = []) {
-    $name = 'templates/' . $name;
-    $result = '';
-
-    if (!is_readable($name)) {
-        return $result;
-    }
-
-    ob_start();
-    extract($data);
-    require $name;
-
-    $result = ob_get_clean();
-
-    return $result;
-}
-
-/**
- * Форматирует цену
- *
- * @param mixed $price
- *
- * @return string|false
- */
-function formatPrice($price)
-{
-    if (!is_numeric($price)) {
-        return false;
-    }
-    return number_format(ceil($price), 0, ',', ' ') . ' ₽';
-}
-
-/**
- * Преобразует специальные символы в строке.
- *
- * @param string $string
- *
- * @return string
- */
-function clearSpecials(string $string)
-{
-    return htmlspecialchars($string);
-}
-
-/**
- * Возвращает период до указанной даты в часах и минутах.
- *
- * @param  string $date Дата из будущего
- *
- * @return string[] Массив, где первый элемент — целое количество часов до даты, а второй — остаток в минутах
- */
-function getTimeUntil(string $date): array
-{
-    $date = date_create($date);
-    if (!$date) {
-        return ['hours' => '00', 'minutes' => '00'];
-    }
-    $diff = date_diff(date_create(), $date);
-    if (date_interval_format($diff, '%r')) {
-        return ['hours' => '00', 'minutes' => '00'];
-    }
-    $days = date_interval_format($diff, '%a');
-    $hours = date_interval_format($diff, '%H');
-    $minutes = date_interval_format($diff, '%I');
-    if ($days) {
-        return ['hours' => $days * 24 + $hours, 'minutes' => $minutes];
-    }
-    return ['hours' => $hours, 'minutes' => $minutes];
-}
-
-/**
- * Отображает сообщение об ошибке.
- * **Завершает выполнение скрипта.**
- *
- * @param string $errMessage Выводимое сообщение
- * @return void
- */
-function showError(string $errMessage): void
-{
-    exit(include_template(
-        'error.php',
-        ['error' => $errMessage]
-    ));
-}
-
-/**
- * Запрашивает категории в БД.
- *
- * @param mysqli $dbConnection Подключение к БД
- * @return array Массив записей
- */
-function getCategories(mysqli $dbConnection): array
-{
-    $sqlQuery = 'SELECT * FROM `categories`';
-    return dbFetchData($dbConnection, $sqlQuery);
-}
-
-/**
- * Запрашивает открытые лоты в БД.
- *
- * @param mysqli $dbConnection Подключение к БД
- * @return array Массив записей
- */
-function getOpenLots(mysqli $dbConnection): array
-{
-    $sqlQuery = 'SELECT l.id, l.title `name`, l.image_path `url`, l.expire_date expiration,
-    IFNULL((SELECT amount FROM bids WHERE lot_id=l.id ORDER BY id DESC LIMIT 1), l.price) price,
-    c.name category
-    FROM lots l JOIN categories c ON l.category_id=c.id
-    WHERE l.expire_date > NOW()
-    ORDER BY l.creation_time DESC, l.id DESC LIMIT 9';
-
-    /*
-    Можно сюда перенести логику обработки оставшегося времени.
-    Добавить элементы массива 'timer' и 'timerclass', например.
-    Тогда в шаблоне будет "чище".
-    */
-
-    return dbFetchData($dbConnection, $sqlQuery);
-}
-
-/**
- * Выполняет запрос к БД и возвращает результат в виде ассоциативного массива.
- *
- * @param mysqli $dbConnection Подключение к БД
- * @param string $sqlQuery Строка запроса
- * @return array Массив записей
- */
-function dbFetchData(mysqli $dbConnection, string $sqlQuery): array
-{
-    $sqlResult = mysqli_query($dbConnection, $sqlQuery);
-
-    if (!$sqlResult) {
-        showError(mysqli_error($dbConnection));
-    }
-
-    return mysqli_fetch_all($sqlResult, MYSQLI_ASSOC);
-}
-
-/**
- * Устанавливает соединение с БД.
- * Принимает массив параметров с ключами 'host', 'user', 'password', 'database'.
- *
- * @param string[] $dbConfig Массив параметров подключения
- * @return mysqli Объект подключения к БД
- */
-function dbConnect($dbConfig): mysqli
-{
-    if (empty($dbConfig)) {
-        showError('Некорректная конфигурация подключения к базе данных.');
-    }
-
-    $dbConnection = mysqli_connect(
-        $dbConfig['host'],
-        $dbConfig['user'],
-        $dbConfig['password'],
-        $dbConfig['database']
-    );
-
-    if (!$dbConnection) {
-        showError(mysqli_connect_error());
-    }
-
-    mysqli_set_charset($dbConnection, 'utf8');
-
-    return $dbConnection;
-}
